fix(web): `sidebar-content` 내부 아이템 props로 주입
